Removed HomeActivity and set up MainActivity
