+++ app/build.gradle.kts
@@ -56,6 +56,7 @@ dependencies {
     implementation(platform("com.google.firebase:firebase-bom:33.4.0"))
     implementation("me.dm7.barcodescanner:zxing:1.9.8")
     implementation ("com.google.zxing:core:3.4.1")
+
     implementation("com.google.firebase:firebase-auth")
     implementation("com.google.firebase:firebase-firestore")
     implementation("com.google.firebase:firebase-storage")
