+++ app/src/main/java/com/example/fusion0/AppNotifications.java
@@ -1,5 +1,6 @@
 package com.example.fusion0;
 
+import android.Manifest;
 import android.app.Activity;
 import android.app.NotificationChannel;
 import android.app.NotificationManager;
@@ -8,18 +9,12 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.os.Build;
-import android.Manifest;
 import android.util.Log;
 
 import androidx.core.app.ActivityCompat;
 import androidx.core.app.NotificationCompat;
 
-import com.google.firebase.firestore.auth.User;
-
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
 
 public class AppNotifications {
     // https://stackoverflow.com/questions/44305206/ask-permission-for-push-notification
