+++ app/src/main/res/layout/activity_favourite.xml
@@ -144,7 +144,7 @@
                 android:layout_weight="1"
                 android:background="@android:color/white"
                 android:contentDescription="search"
-                android:src="@drawable/ic_grey_qrscanner" />
+                android:src="@drawable/ic_grey_camera" />
 
             <ImageButton
                 android:id="@+id/toolbar_add"
