+++ app/src/main/res/values/ic_launcher_background.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="ic_launcher_background">#632B6C</color>
+    <color name="ic_launcher_background">#5763A8</color>
 </resources>
